# backend/driver/openapi.yaml
# Generated with protoc-gen-openapi
# https://github.com/google/gnostic/tree/master/cmd/protoc-gen-openapi

openapi: 3.0.3
info:
    title: Driver API
    version: 0.0.1
paths:
    /api/v1/GetVerifyCode:
        get:
            tags:
                - Driver
            operationId: Driver_GetVerifyCode
            parameters:
                - name: length
                  in: query
                  schema:
                    type: integer
                    format: int32
                - name: type
                  in: query
                  schema:
                    type: integer
                    format: enum
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/GetVerifyCodeReply'
                default:
                    description: Default error response
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Status'
    /api/v1/Login:
        post:
            tags:
                - Driver
            operationId: Driver_Login
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/LoginRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/LoginReply'
                default:
                    description: Default error response
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Status'
    /api/v1/Logout:
        post:
            tags:
                - Driver
            operationId: Driver_Logout
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/LogoutRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/LogoutReply'
                default:
                    description: Default error response
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Status'
    /api/v1/SubmitPhone:
        post:
            tags:
                - Driver
            operationId: Driver_SubmitPhone
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/SubmitPhoneRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/SubmitPhoneReply'
                default:
                    description: Default error response
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Status'
components:
    schemas:
        GetVerifyCodeReply:
            type: object
            properties:
                code:
                    type: integer
                    format: int32
                message:
                    type: string
        GoogleProtobufAny:
            type: object
            properties:
                '@type':
                    type: string
                    description: The type of the serialized message.
            additionalProperties: true
            description: Contains an arbitrary serialized message along with a @type that describes the type of the serialized message.
        LoginReply:
            type: object
            properties:
                code:
                    type: integer
                    format: int32
                message:
                    type: string
                token:
                    type: string
                expireTime:
                    type: string
        LoginRequest:
            type: object
            properties:
                phoneNumber:
                    type: string
                verifyCode:
                    type: string
        LogoutReply:
            type: object
            properties:
                code:
                    type: integer
                    format: int32
                message:
                    type: string
        LogoutRequest:
            type: object
            properties: {}
        Status:
            type: object
            properties:
                code:
                    type: integer
                    description: The status code, which should be an enum value of [google.rpc.Code][google.rpc.Code].
                    format: int32
                message:
                    type: string
                    description: A developer-facing error message, which should be in English. Any user-facing error message should be localized and sent in the [google.rpc.Status.details][google.rpc.Status.details] field, or localized by the client.
                details:
                    type: array
                    items:
                        $ref: '#/components/schemas/GoogleProtobufAny'
                    description: A list of messages that carry the error details.  There is a common set of message types for APIs to use.
            description: 'The `Status` type defines a logical error model that is suitable for different programming environments, including REST APIs and RPC APIs. It is used by [gRPC](https://github.com/grpc). Each `Status` message contains three pieces of data: error code, error message, and error details. You can find out more about this error model and how to work with it in the [API Design Guide](https://cloud.google.com/apis/design/errors).'
        SubmitPhoneReply:
            type: object
            properties:
                code:
                    type: string
                message:
                    type: string
                status:
                    type: string
        SubmitPhoneRequest:
            type: object
            properties:
                phone:
                    type: string
tags:
    - name: Driver